changed int to float
seems there was a decision that promotions are always percent off,
while we are using them as Final Price.

+++ app/app_sitelette/scripts/views/landingSubviews/promotionView.js
@@ -178,7 +178,7 @@ var PromotionView = Backbone.View.extend({
 
   onBuyItem: function(e) {
       var $target = $(e.currentTarget);
-      var promoPrice = parseInt($target.data('price'));
+      var promoPrice = parseFloat($target.data('price'));
       //promoPrice=promoPrice.substr(1, $target.data('price').length - 1) * 100) / 100;
       var promoUUID=$target.data('uuid');
       console.log(" price:"+promoPrice+", uuid:"+promoUUID);
